requirements.txt
redis==2.10.5
python-logstash==0.4.6
requests==2.9.1